# Optimize plain object access