--- a/Jint.Tests.PublicInterface/RavenApiUsageTests.cs
+++ b/Jint.Tests.PublicInterface/RavenApiUsageTests.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using Esprima.Ast;
 using Jint.Constraints;
 using Jint.Native;