--- a/Jint/Native/Prototype.cs
+++ b/Jint/Native/Prototype.cs
@@ -7,7 +7,7 @@ public abstract class Prototype : ObjectInstance
 {
     internal readonly Realm _realm;
 
-    private protected Prototype(Engine engine, Realm realm) : base(engine)
+    private protected Prototype(Engine engine, Realm realm) : base(engine, type: InternalTypes.Object | InternalTypes.PlainObject)
     {
         _realm = realm;
     }