Now with recently added knowledge about object types, we can check if we are accessing plain basic object and skip virtual dispatch and property descriptor renewal when possible.

## Jint.Benchmark.ObjectAccessBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |UpdateObjectProperty|265.8 ms|1.06 ms|90.95 MB|
| **New** |	| **221.9 ms (-17%)** | **0.61 ms** | **60.42 MB (-34%)** |